+++ crates/wdl-format/src/v1/struct.rs
@@ -7,6 +7,7 @@ use crate::PreToken;
 use crate::TokenStream;
 use crate::Writable as _;
 use crate::element::FormatElement;
+use crate::v1::write_sections;
 
 /// Formats a [`StructDefinition`](wdl_ast::v1::StructDefinition).
 ///
@@ -36,24 +37,24 @@ pub fn format_struct_definition(
     stream.end_line();
     stream.increment_indent();
 
-    let mut meta = None;
-    let mut parameter_meta = None;
+    let mut meta_sections = Vec::new();
+    let mut parameter_meta_sections = Vec::new();
     let mut members = Vec::new();
     let mut close_brace = None;
 
     for child in children {
         match child.element().kind() {
             SyntaxKind::MetadataSectionNode => {
-                meta = Some(child.clone());
+                meta_sections.push(child);
             }
             SyntaxKind::ParameterMetadataSectionNode => {
-                parameter_meta = Some(child.clone());
+                parameter_meta_sections.push(child);
             }
             SyntaxKind::UnboundDeclNode => {
-                members.push(child.clone());
+                members.push(child);
             }
             SyntaxKind::CloseBrace => {
-                close_brace = Some(child.clone());
+                close_brace = Some(child);
             }
             _ => {
                 unreachable!(
@@ -64,22 +65,17 @@ pub fn format_struct_definition(
         }
     }
 
-    if let Some(meta) = meta {
-        (&meta).write(stream, config);
-        stream.blank_line();
-    }
-
-    if let Some(parameter_meta) = parameter_meta {
-        (&parameter_meta).write(stream, config);
-        stream.blank_line();
-    }
+    write_sections(&meta_sections, stream, config);
+    write_sections(&parameter_meta_sections, stream, config);
 
     for member in members {
-        (&member).write(stream, config);
+        member.write(stream, config);
     }
 
     stream.decrement_indent();
-    (&close_brace.expect("struct definition close brace")).write(stream, config);
+    close_brace
+        .expect("struct definition close brace")
+        .write(stream, config);
     stream.end_line();
 }